requirements.txt
gitpython
transformers
torch
numpy
faiss-cpu
difflib
pytest
pytest-mock